pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "semantic_find"
version = "0.1.0"
description = "RAG, but for humans! Unlike traditional fuzzy search -- 'California' matching with 'Californuh' -- we'll be building a tool that connects similar concepts: think 'salmon sushi' matching with 'tuna nigiri'."
readme = "README.rst"
authors = [
  {name = "Vincent Selhorst-Jones", email = "vselhorst-jones@chanzuckerberg.com"}
]
maintainers = [
  {name = "Vincent Selhorst-Jones", email = "vselhorst-jones@chanzuckerberg.com"}
]
classifiers = [

]
license = {text = "MIT license"}
dependencies = [
   "psycopg2-binary==2.9.9",
   "numpy", 
   "pgvector"
  
]

[project.optional-dependencies]
dev = [
    "coverage",  # testing
    "mypy",  # linting
    "pytest",  # testing
    "ruff",  # linting
    "flake8" #linting
]

[project.urls]

bugs = "https://github.com/chanzuckerberg/semantic_find/issues"
changelog = "https://github.com/chanzuckerberg/semantic_find/blob/master/changelog.md"
homepage = "https://github.com/chanzuckerberg/semantic_find"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
"*" = ["*.*"]




# Mypy
# ----

[tool.mypy]
files = "."

# Use strict defaults
strict = true
warn_unreachable = true
warn_no_return = true

[[tool.mypy.overrides]]
# Don't require test functions to include types
module = "tests.*"
allow_untyped_defs = true
disable_error_code = "attr-defined"

